requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.9.1
django-crispy-forms==1.6.0
django-filter==0.12.0
django-recaptcha==1.0.5
djangorestframework==3.3.2
gnureadline==6.3.3
ipython==4.1.1
ipython-genutils==0.1.0
Markdown==2.6.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.0
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.1.0
wheel==0.24.0
